+++ lib/commands.ts
@@ -1,4 +1,4 @@
-import { Located } from "./located";
+import { Element } from "./element";
 
 /**
  * TODO: consider moving impl. from Located to here, and then reuse in Located
@@ -8,9 +8,13 @@ export namespace perform {
 
     export const press =
         (text: string, { delay = 0, noWaitAfter = false } = {}) =>
-            (located: Located) => located.press(text, { delay, noWaitAfter })
-            
+            (located: Element) => located.press(text, { delay, noWaitAfter })
+
     export const type =
         (text: string, { delay = 0, noWaitAfter = false } = {}) =>
-            (located: Located) => located.type(text, { delay, noWaitAfter })
+            (located: Element) => located.type(text, { delay, noWaitAfter })
+
+    export const click =
+        (options = {}) => // TODO: consider adding option values here too
+            (located: Element) => located.click(options)
 }