+++ test/acceptance/todomvc.test.ts
@@ -1,4 +1,5 @@
-import { stage, goto, $, perform } from '../../lib'
+import { elements } from './../../lib/playright';
+import { stage, goto, element, perform } from '../../lib'
 
 describe('Todomvc', () => {
 
@@ -8,13 +9,18 @@ describe('Todomvc', () => {
 
     fit('should complete todo', async () => {
 
-        // TODO: slowMo seems to not work... fix!
-        stage.launchOptions = {...stage.launchOptions, slowMo: 1000}; 
+        // TODO: the following somehow does not work... why?
+        // stage.launchOptions = {...stage.launchOptions, slowMo: 1000}; 
 
         await goto('http://todomvc.com/examples/emberjs');
 
-        await $('#new-todo').type('a').then(perform.press('Enter'));
-        await $('#new-todo').type('b').then(perform.press('Enter'));
-        await $('#new-todo').type('c').then(perform.press('Enter'));
+        await element('#new-todo').type('a').then(perform.press('Enter'));
+        await element('#new-todo').type('b').then(perform.press('Enter'));
+        await element('#new-todo').type('c').then(perform.press('Enter'));
+        // await elements('#todo-list li').should();
+
+        await elements('#todo-list li').first.element('.toggle').click();
+
+        await stage.page.waitForTimeout(4000);
     })
 });